Fixed all mistakes

+++ skuratov.danil/P4/countSaddlePoints.cpp
@@ -1,8 +1,10 @@
 #include "countSaddlePoints.h"
 
+#include <cstddef>
+
 int skuratov::countSaddlePoints(int* array, size_t rows, size_t cols)
 {
-  int saddleCount = 0;
+  size_t saddleCount = 0;
   for (size_t i = 0; i < rows; ++i)
   {
     for (size_t j = 0; j < cols; ++j)
